requirements.txt
big-thing-py
pyyaml
paramiko
termcolor
tabulate
rich
